use spring-boot

+++ service/src/main/java/com/epam/esm/exception/ResourceNotFoundException.java
@@ -11,19 +11,19 @@ import lombok.experimental.FieldDefaults;
 @Getter
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class ResourceNotFoundException extends RuntimeException {
-    int idResource;
+    Long idResource;
     String nameResource;
 
     public ResourceNotFoundException(String message) {
         super(message);
     }
 
-    public ResourceNotFoundException(String message, int idResource) {
+    public ResourceNotFoundException(String message, Long idResource) {
         super(message);
         this.idResource = idResource;
     }
 
-    public ResourceNotFoundException(String message, int idResource, String nameResource) {
+    public ResourceNotFoundException(String message, Long idResource, String nameResource) {
         super(message);
         this.idResource = idResource;
         this.nameResource = nameResource;
